==> 03 Data description.qmd <==
---
title: "Data Description"
author: "Eldin, Ahmed, Kelly, Miguel*"
format: html
editor: visual
---

## Quarto

Quarto enables you to weave together content and executable code into a finished document. To learn more about Quarto see <https://quarto.org>.

```{r}
data_description <- readxl::read_excel("Data/Data.xlsx", sheet = "Table S4", range = "A1:A1", col_names = FALSE)

#Description of variables 

data_description <- as.data.frame(data_description)
data_description[1,1]
```

`r data_description[1,1]`
